refactor: Move services into separate folders
